Use checkStyle to detect coding style violations
