+++ src/main/java/Launcher.java
@@ -1,6 +1,9 @@
-import Icarus.Main;
+import icarus.Main;
 import javafx.application.Application;
 
+/**
+ * TODO
+ */
 public class Launcher {
     public static void main(String[] args) {
         Application.launch(Main.class, args);
