+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ---------------------------------------------
-Greetings! My name is Icarus.Icarus.
+Greetings! My name is icarus.icarus.
 How can I be of service to you today?
 ---------------------------------------------
 ---------------------------------------------
